Переопределение операндов для структуры; более внятный режим ToString() для комплексных чисел и дробей.
